+++ p2p/dandelion/src/tests/mod.rs
@@ -1,5 +1,5 @@
+mod pool;
 mod router;
-mod txpool;
 
 use std::{collections::HashMap, future::Future, hash::Hash, sync::Arc};
